Move fixture related classes alongside driver classes and away from tests
