+++ src/test/java/example/Example.java
@@ -1,5 +1,7 @@
 package example;
 
+import org.concordion.cubano.template.fixture.CubanoDemoIndex;
+
 /**
  * Run me as a Junit Test to test all specifications.
  */
