+++ src/test/java/example/commands/AssertEqualsFixture.java
@@ -1,6 +1,6 @@
 package example.commands;
 
-import example.CubanoDemoFixture;
+import org.concordion.cubano.template.fixture.CubanoDemoFixture;
 
 public class AssertEqualsFixture extends CubanoDemoFixture {
