+++ src/test/java/example/commands/Commands.java
@@ -1,6 +1,6 @@
 package example.commands;
 
-import example.CubanoDemoIndex;
+import org.concordion.cubano.template.fixture.CubanoDemoIndex;
 
 public class Commands extends CubanoDemoIndex {
