+++ src/test/java/example/concordion/Concordion.java
@@ -1,6 +1,6 @@
 package example.concordion;
 
-import example.CubanoDemoIndex;
+import org.concordion.cubano.template.fixture.CubanoDemoIndex;
 
 public class Concordion extends CubanoDemoIndex {
