+++ src/test/java/example/cubanoconcordion/ExceptionHtmlCaptureExtensionFixture.java
@@ -1,7 +1,7 @@
 package example.cubanoconcordion;
 
-import example.CubanoDemoBrowserFixture;
 import org.concordion.cubano.template.driver.ui.google.GoogleSearchPage;
+import org.concordion.cubano.template.fixture.CubanoDemoBrowserFixture;
 
 public class ExceptionHtmlCaptureExtensionFixture extends CubanoDemoBrowserFixture {
