+++ src/test/java/example/cubanocore/DataLoadersFixture.java
@@ -2,10 +2,11 @@ package example.cubanocore;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import example.CubanoDemoFixture;
+
 import org.concordion.cubano.data.JsonLoader;
 import org.concordion.cubano.driver.http.JsonReader;
 import org.concordion.cubano.template.driver.domain.User;
+import org.concordion.cubano.template.fixture.CubanoDemoFixture;
 import org.concordion.ext.StoryboardMarkerFactory;
 import org.concordion.ext.storyboard.CardResult;
 import org.concordion.ext.storyboard.StockCardImage;
