+++ src/test/java/example/cubanocore/DataManagementFixture.java
@@ -11,8 +11,7 @@ import org.concordion.cubano.framework.resource.ResourceScope;
 import org.concordion.cubano.template.driver.database.DataManagementDatabase;
 import org.concordion.cubano.template.driver.domain.AbcDomainData;
 import org.concordion.cubano.template.driver.services.AbcService;
-
-import example.CubanoDemoFixture;
+import org.concordion.cubano.template.fixture.CubanoDemoFixture;
 
 public class DataManagementFixture extends CubanoDemoFixture {
