+++ src/test/java/example/cubanowebdriver/CubanoWebDriver.java
@@ -1,6 +1,6 @@
 package example.cubanowebdriver;
 
-import example.CubanoDemoIndex;
+import org.concordion.cubano.template.fixture.CubanoDemoIndex;
 
 public class CubanoWebDriver extends CubanoDemoIndex {
