+++ src/test/java/example/cubanowebdriver/SwitchBrowser.java
@@ -1,9 +1,9 @@
 package example.cubanowebdriver;
 
-import example.CubanoDemoBrowserFixture;
 import org.concordion.cubano.driver.web.Browser;
 import org.concordion.cubano.driver.web.provider.ChromeBrowserProvider;
 import org.concordion.cubano.template.driver.ui.google.GoogleSearchPage;
+import org.concordion.cubano.template.fixture.CubanoDemoBrowserFixture;
 
 import java.io.IOException;
