+++ src/test/java/org/concordion/cubano/template/fixture/CubanoDemoBrowserFixture.java
@@ -1,4 +1,4 @@
-package example;
+package org.concordion.cubano.template.fixture;
 
 import java.io.Closeable;
